more paranoid tests.py

+++ tests.py
@@ -63,7 +63,10 @@ def dump_output(basedir, compl, stdoutname, stderrname, imagename, referencename
         if CAPTURE_STDERR:
             with open(stderrname, "w") as outfile:
                 outfile.write(compl.stderr.decode('utf-8'))
-    os.replace(RESULT_NAME, imagename)
+    if file.path.exists(RESULT_NAME):
+        os.replace(RESULT_NAME, imagename)
+    else:
+        print("no result image")
 
     reportstring += f"""
 <button class="test coloredblock{' failed' if not tr.passed else ''}">{tr.testfile}: {tr.result}</button>
